+++ kinto/__main__.py
@@ -2,7 +2,8 @@ from __future__ import print_function
 import argparse
 import os
 import sys
-
+import logging
+import logging.config
 from six.moves import input
 from kinto.core import scripts
 from pyramid.scripts import pserve
@@ -44,8 +45,14 @@ def main(args=None):
     parser_init.set_defaults(which='init')
 
     parser_migrate = subparsers.add_parser('migrate')
+    parser_migrate.add_argument('--dry-run',
+                                action='store_true',
+                                help='Simulate the migration operations '
+                                     'and show information',
+                                dest='dry_run',
+                                required=False,
+                                default=False)
     parser_migrate.set_defaults(which='migrate')
-
     parser_start = subparsers.add_parser('start')
     parser_start.add_argument('--reload',
                               action='store_true',
@@ -92,8 +99,10 @@ def main(args=None):
                 pip.main(['install', "kinto[postgresql]"])
 
     elif parsed_args['which'] == 'migrate':
+        dry_run = parsed_args['dry_run']
+        logging.config.fileConfig(config_file)
         env = bootstrap(config_file)
-        scripts.migrate(env)
+        scripts.migrate(env, dry_run=dry_run)
 
     elif parsed_args['which'] == 'start':
         pserve_argv = ['pserve', config_file]