+++ kinto/core/cache/redis.py
@@ -37,7 +37,7 @@ class Cache(CacheBase):
     def settings(self):
         return dict(self._client.connection_pool.connection_kwargs)
 
-    def initialize_schema(self):
+    def initialize_schema(self, dry_run=False):
         # Nothing to do.
         pass