+++ kinto/core/permission/postgresql/__init__.py
@@ -6,6 +6,7 @@ from collections import OrderedDict
 
 from kinto.core import logger
 from kinto.core.permission import PermissionBase
+from kinto.core.storage.exceptions import BackendError
 from kinto.core.storage.postgresql.client import create_from_config
 
 
@@ -66,11 +67,24 @@ class Permission(PermissionBase):
         super(Permission, self).__init__(*args, **kwargs)
         self.client = client
 
-    def initialize_schema(self):
+    def initialize_schema(self, dry_run=False):
+        try:
+            self.get_user_principals("TRYING")
+            logger.info("PostgreSQL permission schema is up-to-date.")
+            return
+        except BackendError:
+            pass
+
         # Create schema
         here = os.path.abspath(os.path.dirname(__file__))
-        schema = open(os.path.join(here, 'schema.sql')).read()
+        sql_file = os.path.join(here, 'schema.sql')
+
+        if dry_run:
+            logger.info("Create permission schema from %s" % sql_file)
+            return
+
         # Since called outside request, force commit.
+        schema = open(sql_file).read()
         with self.client.connect(force_commit=True) as conn:
             conn.execute(schema)
         logger.info('Created PostgreSQL permission tables')